worker-service/requirements.txt
fastapi
uvicorn
redis
python-dotenv
# We will add playwright and langchain later in Phase 4